Update TO-DO.txt
ok so i tried to fix some of these issues, but it only ended up causing other problems.
if we redraw the canvas before drawing the screenshot, the client loses a lot of the strokes, probably due to timing. while writing this, another idea came to mind ill try that next.
Ik how to solve some issues like drawing outside canvas and resetting canvas, and i think the subtle duplication shouldnt be too bad either.
The MAIN issue is the same as the one we were having before, with taking an image of the screen rather than the window.